chore: Rename example tests from internal packages into a new module
